fixed major game logic bug on growth algo; building.update fixed
