+++ src/main/java/edu/mscd/thesis/model/bldgs/Home.java
@@ -16,24 +16,7 @@ public abstract class Home extends AbstractBuilding {
 		super(pos, tileType, zoneType, density);
 	}
 
-	@Override
-	public double update(double growthValue) {
-		int tileMaxDensity = this.getTileType().getMaxDensity().getDensityLevel();
-		int currentDensityLevel = this.getDensity().getDensityLevel();
-		int currentOccupancy = this.currentOccupancy();
-		int maxOccupancy = this.getMaxOccupants();
-		if(currentOccupancy>=maxOccupancy && growthValue>Rules.GROWTH_THRESHOLD && tileMaxDensity>currentDensityLevel){
-			this.changeDensity(getDensity().getNextLevel());
-			return growthValue - Rules.BASE_GROWTH_COST;
-		}else if(currentDensityLevel==0 && growthValue>Rules.GROWTH_THRESHOLD){
-			this.changeDensity(getDensity().getNextLevel());
-			return growthValue - Rules.BASE_GROWTH_COST;
-		}else if(growthValue>Rules.GROWTH_THRESHOLD){
-			this.changeDensity(getDensity().getPrevLevel());
-			return growthValue + Rules.BASE_GROWTH_COST*2;
-		}
-		return growthValue;
-	}
+
 
 	@Override
 	public void setMaxOccupancy(int max) {
